v3.0.1
- Implemented a new method for using ephemeral messages.
- Added an unban command for the dev category commands.
-Renamed nodeClose.js to nodeClosed.js to align with Rainlink node events.
-Updated the song search logic in the play command to match the latest Rainlink version.

Added a default volume option when creating a player.

+++ src/events/rainlink/player/queueEmpty.js
@@ -1,6 +1,8 @@
 const { EmbedBuilder } = require("discord.js");
 
 module.exports = async (client, player) => {
+    if (!player) return;
+
     if (player.message) player.message.delete().catch((e) => {});
 
     const channel = await client.channels.cache.get(player.textId);
